backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
groq>=0.4.0
python-dotenv>=1.0.1
pydantic>=2.10.0
